소스정리 / css/js/ejs 폴더 만들어서 모듈 별로 따로 정리 common.js 추가해서 기능모듈은 여기서 import 됨. ejs에서 메타태그, favicon 태그를 따로 빼서 모듈화 / channelCreate.ejs, channelJoin.ejs 방 이름, 방 패스워드 추가
