+++ public/css/channel/remoteDisplay.css
