+++ public/css/mainPage.css
@@ -1,8 +1,9 @@
 @import url("https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600&display=swap");
 @import url("https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css");
-@import "./leftnavbar.css";
-@import "./bodyContainer.css";
-@import "./bootstrap.min.css";
+@import "./mainpage/leftnavbar.css";
+@import "./mainpage/bodyContainer.css";
+@import "./mainpage/channelCreate.css";
+@import "./lib/bootstrap.min.css";
 
 * {
     margin : 0;
