+++ public/css/mainpage/channelCreate.css
@@ -1,0 +1,3 @@
+.roomSetting-info-group {
+    flex-direction: row-reverse;
+}
