+++ views/headparts/meta-parts.ejs
@@ -1,0 +1,14 @@
+<!-- meta tag -->
+<meta charset="UTF-8">
+<meta http-equiv="X-UA-Compatible" content="IE=edge">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta name="description"
+    content="X-Frame-Bypass: Web Component extending IFrame to bypass X-Frame-Options: deny/sameorigin">
+<meta name="description" content="Owake Me, 내가 원하는 곳 어디에서나 시간을 초월한 세계로 나를 연결하는 거대한 네트워크!" />
+<meta name="description"
+    content="Owake.Me, A huge network connecting me into the timeless world wherever I want. 
+Kron World, You govern your world, Proof of Moment, Proof of World Login, Time, Timeless in Metaverse Environment, Kron Economy" />
+<meta name="robots" />
+
+<meta name="google" content="notranslate" />
+<!-- meta tag end -->
