+++ views/jsparts/room-parts.ejs
@@ -1,0 +1,13 @@
+<script src="rtcClient.js"></script>
+<script src="rtmClient.js" type="module"></script>
+<script src="common.js" type="module"></script>
+<!-- <script src="f12defense.js"></script> -->
+<!-- <script src="recodingDeviceCtrl.js"></script> -->
+<!-- <script src="momentShare.js"></script> -->
+<!-- <script src="screenShare.js"></script> -->
+<!-- <script src="fileShare.js"></script> -->
+<!-- <script src="muteUtils.js"></script> -->
+<!-- <script src="copyUrl.js"></script> -->
+<!-- <script src="chat.js"></script> -->
+<!-- <script src="remoteDisplay.js"></script> -->
+<!-- <script src="options-btn.js"></script> -->
