+++ views/popup/channelCreate.ejs
@@ -32,9 +32,18 @@
                         <div class="col-sm">
                             <input id="private_active" name="drone" type="radio" value="Private Channel" />
                             <label for="private_active">Private</label>
-                            <input type="text">
+                        </div>
+
+                        <div class="col-sm">
+                            <p class="channel-info-text">Channel Name</p>
+                            <input type="text" placeholder="Enter the Channel Name">
+                        </div>
+                        <div class="col-sm">
+                            <p class="channel-info-text">Channel Password</p>
+                            <input type="text" placeholder="Enter the Channel Password">
                         </div>
                     </div>
+                    <br><br>
                     <div class="row category-info-group">
 
                         <div class="col-sm">
