+++ views/popup/channelJoin.ejs
@@ -21,6 +21,15 @@
                             <p class="admin-info-text">User Password</p>
                             <input id="password" type="text" placeholder="Enter the Password" />
                         </div>
+
+                        <div class="col-sm" style="margin-top: 10px;">
+                            <p class="channel-info-text">Channel Name</p>
+                            <input id="password" type="text" placeholder="Enter the Channel Name" />
+                        </div>
+                        <div class="col-sm" style="margin-top: 10px;">
+                            <p class="channel-info-text">Channel Password</p>
+                            <input id="password" type="text" placeholder="Enter the Channel Password" />
+                        </div>
                     </div>
                     <br><br>
                     <div class="button-group">
