+++ views/room.ejs
@@ -2,43 +2,8 @@
 <html lang="en">
 
 <head>
-    <!-- meta tag -->
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description"
-        content="X-Frame-Bypass: Web Component extending IFrame to bypass X-Frame-Options: deny/sameorigin">
-    <meta name="description" content="Owake Me, 내가 원하는 곳 어디에서나 시간을 초월한 세계로 나를 연결하는 거대한 네트워크!" />
-    <meta name="description"
-        content="Owake.Me, A huge network connecting me into the timeless world wherever I want. 
-    Kron World, You govern your world, Proof of Moment, Proof of World Login, Time, Timeless in Metaverse Environment, Kron Economy" />
-    <meta name="robots" />
 
-    <meta name="google" content="notranslate" />
-    <!-- meta tag end -->
-
-    <!-- favicon package -->
-    <link rel="shortcut icon" href="favicon.ico">
-    <link rel="apple-touch-icon" sizes="57x57" href="apple-icon-57x57.png">
-    <link rel="apple-touch-icon" sizes="60x60" href="apple-icon-60x60.png">
-    <link rel="apple-touch-icon" sizes="72x72" href="apple-icon-72x72.png">
-    <link rel="apple-touch-icon" sizes="76x76" href="apple-icon-76x76.png">
-    <link rel="apple-touch-icon" sizes="114x114" href="apple-icon-114x114.png">
-    <link rel="apple-touch-icon" sizes="120x120" href="apple-icon-120x120.png">
-    <link rel="apple-touch-icon" sizes="144x144" href="apple-icon-144x144.png">
-    <link rel="apple-touch-icon" sizes="152x152" href="apple-icon-152x152.png">
-    <link rel="apple-touch-icon" sizes="180x180" href="apple-icon-180x180.png">
-    <link rel="icon" type="image/png" sizes="192x192" href="android-icon-192x192.png">
-    <link rel="icon" type="image/png" sizes="32x32" href="favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="96x96" href="favicon-96x96.png">
-    <link rel="icon" type="image/png" sizes="16x16" href="favicon-16x16.png">
-    <link rel="manifest" href="manifest.json">
-    <meta name="msapplication-TileColor" content="#ffffff">
-    <meta name="msapplication-TileImage" content="ms-icon-144x144.png">
-    <meta name="theme-color" content="#ffffff">
-    <!-- favicon End -->
-
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="channel.css">
 
     <script src="owake_rtc_n.js"></script>
     <script src="owake_rtm.js"></script>
@@ -145,21 +110,8 @@
             </div>
         </div>
     </div>
-
-    <script src="f12defense.js"></script>
-    <script src="rtcClient.js"></script>
-    <script src="rtmClient.js" type="module"></script>
-
-    <script src="recodingDeviceCtrl.js"></script>
-    <script src="momentShare.js"></script>
-    <script src="screenShare.js"></script>
-    <script src="fileShare.js"></script>
-    <script src="muteUtils.js"></script>
-    <!-- <script src="copyUrl.js"></script> -->
-    <script src="chat.js"></script>
-    <script src="remoteDisplay.js"></script>
-    <script src="options-btn.js"></script>
-
 </body>
+<!-- Script Import -->
+<%- include('./jsparts/room-parts') %>
 
 </html>
